command_ref doc: fix "–userinstalled"
I noticed this being rendered as "–userinstalled" on Fedora 27 with dnf-2.7.5-1.fc27.noarch.
It should be "--userinstalled" instead.

It's surprising as the same text is fine in Fedora 26 with dnf-2.7.5-1.fc26.noarch.
I think it is very simple to "work around" this though.

I found the sentence confusing as well.  I have attempted to reword as necessary to avoid my own confusion.
Something about using "but" in this context made it sound as if it was about to explain
why the "repoquery" command is _not_ as powerful.  When what it actually goes on to say
is the exact opposite.